+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/apirequest/generator/metric/protocol/ProtocolChain.java
@@ -65,6 +65,7 @@ public class ProtocolChain {
     )
             throws UnknownProtocolValueException {
         LogicalMetric soFar = fromMetric;
+
         for (Protocol p: protocols) {
             if (! apiMetric.getParameters().containsKey(p.getCoreParameterName()))  {
                 continue;
@@ -74,8 +75,8 @@ public class ProtocolChain {
                 break;
             }
 
-            ProtocolMetric soFarProtocol = (ProtocolMetric) soFar;
-            if (!soFarProtocol.accepts(p.getContractName())) {
+            ProtocolMetric soFarProtocolMetric = (ProtocolMetric) soFar;
+            if (!soFarProtocolMetric.accepts(p.getContractName())) {
                 String message = "Protocol triggering parameter is sent on incompatible protocol.";
                 if (strictValidation) {
                     throw new IllegalArgumentException(message);
@@ -85,7 +86,7 @@ public class ProtocolChain {
                 }
             }
 
-            soFar = soFarProtocol.accept(resultMetadata, p.getContractName(), apiMetric.getParameters());
+            soFar = soFarProtocolMetric.accept(resultMetadata, p.getContractName(), apiMetric.getParameters());
         }
         return soFar;
     }